+++ main.go
@@ -106,7 +106,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "1.9.0-alpha.3"
+var ReleaseVersion = "1.9.0-alpha.4-dev"
 
 func main() {
 	exitCode := main2()